Replace `LazyColumn` with `Column` in `TeiDetailDashboard`
`LazyColumn` sets infinite height, if we have to place it inside another scroll view Compose will throw and runtime error saying that. We need to either set a fixed height to the `LazyColumn` or replace it with `Column` which wraps the content. Second one seemed like a good idea, so that it can react to content changes and recalculate the height.