fix: [ANDROAPP-5663] Tei dashboard event list scrolling getting clipped (#3433)
* Replace `LazyColumn` with `Column` in `TeiDetailDashboard`

`LazyColumn` sets infinite height, if we have to place it inside another scroll view Compose will throw and runtime error saying that. We need to either set a fixed height to the `LazyColumn` or replace it with `Column` which wraps the content. Second one seemed like a good idea, so that it can react to content changes and recalculate the height.

* Wrap `cardLayout` and `tei_recycler` in nested scroll view

We have to 2 options to make the entire content scroll in this screen.

1. We migrate the events items to use Compose as well and move them into `TeiDetailDashboard`. Which is ideal, but requires significant migration of UI and associated code with it. Since that's out of scope didn't go with that.

2. Since in earlier commit I have change the `LazyColumn` to `Column`, we can now place the card and the events list inside a nested scroll view to make the entire content scroll.

The 2nd option seems like a sensible change for now, until we migrate events list and eventually this entire screen.

* Add new tei dashboard card support for landscape view

* Disable over scroll mode for nested scroll view and recycler view

* Show `card_front` in TEI dashboard landscape view only if program is available