feat: improve PeriodSelector UX when future periods are enabled (#4371)
When selecting a time period in `DataSetPeriodDialog` for a data set with `openFuturePeriods > 0` the user was forced to scroll down to "today's" period e.g. today, this week, this month and so on. For data sets with a large number of `openFuturePeriods` this could take quite some time.

This change introduces a launch effect which, when `openFuturePeriods` is > 0, will automatically scroll to the time period corresponding to "today" e.g. today, this week, this month or this year. This removes the need to scroll so much.

If the user wants to enter a future period they can simply scroll upwards in the period selector.